+++ website/src/routes/api/(types)/LooseTupleIssue/properties.ts
@@ -1,0 +1,31 @@
+import type { PropertyProps } from '~/components';
+
+export const properties: Record<string, PropertyProps> = {
+  BaseIssue: {
+    modifier: 'extends',
+    type: {
+      type: 'custom',
+      name: 'BaseIssue',
+      href: '../BaseIssue/',
+      generics: ['unknown'],
+    },
+  },
+  kind: {
+    type: {
+      type: 'string',
+      value: 'schema',
+    },
+  },
+  type: {
+    type: {
+      type: 'string',
+      value: 'loose_tuple',
+    },
+  },
+  expected: {
+    type: {
+      type: 'string',
+      value: 'Array',
+    },
+  },
+};