+++ website/src/routes/api/(types)/TupleSchema/properties.ts
@@ -22,6 +22,7 @@ export const properties: Record<string, PropertyProps> = {
             {
               type: 'custom',
               name: 'TupleIssue',
+              href: '../TupleIssue/',
             },
           ],
         },